requirements.txt
annotated-types==0.7.0
anyio==4.4.0
beautifulsoup4==4.12.3
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
fastcore==1.6.2
fastlite==0.0.7
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
numpy==2.0.1
oauthlib==3.2.2
packaging==24.1
pandas==2.2.2
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-fasthtml==0.2.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlite_minutils==3.36.0.post4
starlette==0.38.2
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.5
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
